+++ examples/zustand-widget-config/src/store/useWidgetConfig.ts
@@ -1,0 +1,11 @@
+import { useWidgetConfigStore } from './createWidgetConfigStore.ts';
+
+export const useWidgetConfig = () => {
+  const { config } = useWidgetConfigStore((state) => ({
+    config: state.config,
+  }));
+
+  return {
+    config,
+  };
+};