+++ packages/widget/src/components/AppContainer.tsx
@@ -1,10 +1,18 @@
 import { Box, Container, ScopedCssBaseline, styled } from '@mui/material'
 import type { PropsWithChildren } from 'react'
 import { defaultMaxHeight } from '../config/constants.js'
-import { useHeaderHeight } from '../hooks/useHeaderHeight.js'
 import { useWidgetConfig } from '../providers/WidgetProvider/WidgetProvider.js'
+import { useHeaderHeight } from '../stores/header/useHeaderStore.js'
 import type { WidgetVariant } from '../types/widget.js'
 import { ElementId, createElementId } from '../utils/elements.js'
+
+// NOTE: the setting of the height in AppExpandedContainer, RelativeContainer and CssBaselineContainer can
+//  be done dynamically by values in the config - namely the config.theme.container values display, maxHeight and height
+//  A Number of other components and hooks work with height values that are often set on or derived from these elements
+//  if there are changes to how the height works here you should also check the functionality of these hooks and their point of use
+//    - useTokenListHeight
+//    - useSetContentHeight
+//  Also check any code that is using the methods from elements.ts utils file
 
 export const AppExpandedContainer = styled(Box, {
   shouldForwardProp: (prop) => prop !== 'variant',
@@ -63,34 +71,29 @@ interface CssBaselineContainerProps {
 const CssBaselineContainer = styled(ScopedCssBaseline, {
   shouldForwardProp: (prop) =>
     !['variant', 'paddingTopAdjustment', 'elementId'].includes(prop as string),
-})<CssBaselineContainerProps>(
-  ({ theme, variant, paddingTopAdjustment, elementId }) => ({
-    display: 'flex',
-    flex: 1,
-    flexDirection: 'column',
-    overflowX: 'clip',
-    margin: 0,
-    width: '100%',
-    maxHeight:
-      variant === 'drawer' || theme.container?.display === 'flex'
-        ? 'none'
-        : theme.container?.maxHeight
-          ? theme.container?.maxHeight
-          : theme.container?.height || defaultMaxHeight,
-    overflowY: 'auto',
-    height: theme.container?.display === 'flex' ? 'auto' : '100%',
-    paddingTop: paddingTopAdjustment,
-    // The following allows for the token list to always fill the available height
-    // use of CSS.escape here is to deal with characters such as ':' returned by reacts useId hook
-    //   related issue - https://github.com/facebook/react/issues/26839
-    [`&:has(#${CSS.escape(createElementId(ElementId.TokenList, elementId))})`]:
-      {
-        height: theme.container?.maxHeight
-          ? theme.container?.maxHeight
-          : theme.container?.height || defaultMaxHeight,
-      },
-  })
-)
+})<CssBaselineContainerProps>(({ theme, variant, paddingTopAdjustment }) => ({
+  display: 'flex',
+  flex: 1,
+  flexDirection: 'column',
+  overflowX: 'clip',
+  margin: 0,
+  width: '100%',
+  maxHeight:
+    variant === 'drawer' || theme.container?.display === 'flex'
+      ? 'none'
+      : theme.container?.maxHeight
+        ? theme.container?.maxHeight
+        : theme.container?.height || defaultMaxHeight,
+  overflowY: 'auto',
+  height: theme.container?.display === 'flex' ? 'auto' : '100%',
+  paddingTop: paddingTopAdjustment,
+  // This allows FullPageContainer.tsx to expand and fill the available vertical space in max height and default layout modes
+  '&:has(.full-page-container)': {
+    height: theme.container?.maxHeight
+      ? theme.container?.maxHeight
+      : theme.container?.height || defaultMaxHeight,
+  },
+}))
 
 export const FlexContainer = styled(Container)({
   display: 'flex',