+++ packages/widget/src/components/Header/Header.tsx
@@ -1,7 +1,8 @@
 import type { FC, PropsWithChildren } from 'react'
+import { useLayoutEffect, useRef } from 'react'
 import { useLocation } from 'react-router-dom'
 import { useDefaultElementId } from '../../hooks/useDefaultElementId.js'
-import { useHeaderHeight } from '../../hooks/useHeaderHeight.js'
+import { useSetHeaderHeight } from '../../stores/header/useHeaderStore.js'
 import { ElementId, createElementId } from '../../utils/elements.js'
 import { stickyHeaderRoutes } from '../../utils/navigationRoutes.js'
 import { Container } from './Header.style.js'
@@ -11,13 +12,37 @@ import { WalletHeader } from './WalletHeader.js'
 export const HeaderContainer: FC<PropsWithChildren> = ({ children }) => {
   const { pathname } = useLocation()
   const elementId = useDefaultElementId()
-  const { headerHeight } = useHeaderHeight()
+  const headerRef = useRef<HTMLDivElement>(null)
+  const { setHeaderHeight } = useSetHeaderHeight()
+
+  useLayoutEffect(() => {
+    const handleHeaderResize = () => {
+      const height = headerRef.current?.getBoundingClientRect().height
+
+      if (height) {
+        setHeaderHeight(height)
+      }
+    }
+
+    let resizeObserver: ResizeObserver
+
+    if (headerRef.current) {
+      resizeObserver = new ResizeObserver(handleHeaderResize)
+      resizeObserver.observe(headerRef.current)
+    }
+
+    return () => {
+      if (resizeObserver) {
+        resizeObserver.disconnect()
+      }
+    }
+  }, [setHeaderHeight])
 
   return (
     <Container
       id={createElementId(ElementId.Header, elementId)}
       sticky={stickyHeaderRoutes.some((route) => pathname.includes(route))}
-      maxHeight={headerHeight}
+      ref={headerRef}
     >
       {children}
     </Container>