+++ packages/widget/src/pages/SelectTokenPage/SearchTokenInput.tsx
@@ -1,11 +1,8 @@
-import { Search } from '@mui/icons-material'
-import { FormControl, InputAdornment } from '@mui/material'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
-import { InputCard } from '../../components/Card/InputCard.js'
+import { SearchInput } from '../../components/Search/SearchInput.js'
 import { useFieldActions } from '../../stores/form/useFieldActions.js'
 import { useFieldController } from '../../stores/form/useFieldController.js'
-import { Input } from './SearchTokenInput.style.js'
 
 export const SearchTokenInput = () => {
   const { t } = useTranslation()
@@ -22,27 +19,12 @@ export const SearchTokenInput = () => {
   )
 
   return (
-    <InputCard>
-      <FormControl fullWidth>
-        <Input
-          size="small"
-          placeholder={t('main.tokenSearch') as string}
-          endAdornment={
-            <InputAdornment position="end">
-              <Search />
-            </InputAdornment>
-          }
-          inputProps={{
-            inputMode: 'search',
-            onChange: (e) => onChange((e.target as HTMLInputElement).value),
-            onBlur,
-            name,
-            value,
-            maxLength: 128,
-          }}
-          autoComplete="off"
-        />
-      </FormControl>
-    </InputCard>
+    <SearchInput
+      name={name}
+      placeholder={t('main.tokenSearch')}
+      onChange={(e) => onChange((e.target as HTMLInputElement).value)}
+      onBlur={onBlur}
+      value={value as string | undefined}
+    />
   )
 }