+++ packages/widget/src/pages/SelectTokenPage/useTokenListHeight.ts
@@ -2,19 +2,20 @@ import { debounce, useTheme } from '@mui/material'
 import type { MutableRefObject } from 'react'
 import { useLayoutEffect, useState } from 'react'
 import { useDefaultElementId } from '../../hooks/useDefaultElementId.js'
-import { ElementId, createElementId } from '../../utils/elements.js'
+import {
+  ElementId,
+  getAppContainer,
+  getHeaderElement,
+  getScrollableContainer,
+} from '../../utils/elements.js'
 
 const getContentHeight = (
   elementId: string,
   listParentRef: MutableRefObject<HTMLUListElement | null>
 ) => {
-  const containerElement = document.getElementById(
-    createElementId(ElementId.ScrollableContainer, elementId)
-  )
+  const containerElement = getScrollableContainer(elementId)
 
-  const headerElement = document.getElementById(
-    createElementId(ElementId.Header, elementId)
-  )
+  const headerElement = getHeaderElement(elementId)
 
   const listParentElement = listParentRef?.current
 
@@ -53,6 +54,10 @@ interface UseContentHeightProps {
 export const minTokenListHeight = 360
 export const minMobileTokenListHeight = 160
 
+// NOTE: this hook is implicitly tied to the widget height functionality in the
+//   AppExpandedContainer, RelativeContainer and CssBaselineContainer components as defined in AppContainer.ts
+//   CSS changes in those components can have implications for the functionality in this hook
+
 export const useTokenListHeight = ({
   listParentRef,
   headerRef,
@@ -71,9 +76,7 @@ export const useTokenListHeight = ({
     // calling this on initial mount prevents the lists resizing appearing glitchy
     handleResize()
 
-    const appContainer = document.getElementById(
-      createElementId(ElementId.AppExpandedContainer, elementId)
-    )
+    const appContainer = getAppContainer(elementId)
 
     let resizeObserver: ResizeObserver
     if (appContainer) {