+++ packages/widget/src/types/events.ts
@@ -1,4 +1,5 @@
 import type { ChainId, ChainType, Process, Route } from '@lifi/sdk'
+import type { NavigationRouteType } from '../utils/navigationRoutes.js'
 
 export enum WidgetEvent {
   RouteExecutionStarted = 'routeExecutionStarted',
@@ -11,12 +12,16 @@ export enum WidgetEvent {
   SourceChainTokenSelected = 'sourceChainTokenSelected',
   DestinationChainTokenSelected = 'destinationChainTokenSelected',
   SendToWalletToggled = 'sendToWalletToggled',
+  /**
+   * @deprecated Use `PageEntered` event instead.
+   */
   ReviewTransactionPageEntered = 'reviewTransactionPageEntered',
   /**
    * @deprecated use useWalletManagementEvents hook.
    */
   WalletConnected = 'walletConnected',
   WidgetExpanded = 'widgetExpanded',
+  PageEntered = 'pageEntered',
 }
 
 export type WidgetEvents = {
@@ -36,6 +41,7 @@ export type WidgetEvents = {
    */
   walletConnected: WalletConnected
   widgetExpanded: boolean
+  pageEntered: NavigationRouteType
 }
 
 export interface ContactSupport {